Head *and* Tail NEW-LINE positions, APPEND/LINE (+INSERT,CHANGE)
Historically each value cell in a Rebol array offered a place to put
a "new line marker".  This new line marker would indicate that a
new-line should be output *before* that value was molded.

Yet an N-element array has N + 1 "in-between" positions where a newline
might be indicated.  This approach meant that there was no way to
say that the tail of an array had a new line on it...meaning that one
could not write:

    data: copy [
        a b c
    ]
    new-line (tail data) true
    append data 10

...and get back:

    [
        a b c
        10
    ]

Because there was no memory in C of the fact that a new-line was at
the tail.  The only reason it appears there is a new line is that
there was some logic in mold that said if any new lines were output
during a mold, then the close brace would be on its own line.  You
could not preserve the absence, e.g. with:

    data: copy [
        a b c]

The scanner would lose the distinction, and wind up molding that out
with the close bracket on its own line.

This commit gets the extra "+ 1" bit from the array itself, denoting
if the tail of the array has a new-line bit.  When arrays are merged
then the product takes the union of the new-line bits for the ends:
if a new-line exists on either the head or tail of an insertion it
will update the relevant bits in the target array cells or the array's
tail newline bit.

To make it easier to work with these features, this also adds a /LINE
refinement to APPEND.  This will ensure a newline bit is put on the
tail of the inserted material.  As an added heurstic for /LINE, any
array that does not have a newline bit on its head will be changed to
have one, so one need not write:

    x: copy []
    new-line x true ;-- not needed, will happen implicitly
    append/line x [line one]
    append/line x [line two]