Fix ACTION! values in function default parameters via UNEVAL
UNEVAL was not being called when a default argument was being
evaluated, so it tried to run the function...which isn't what you
would want if you were attempting to capture a function into a
frame variable.  This adds that, as well as not bothering to use the
DEFAULT function when the argument is a local (as it is known to not
be holding a value).

Includes a few other stray changes...fix for a PROTECT crash, fix
for DUMP-VAL on not set arguments, adds `composeII` to try it as
a name, and adds a couple simple tests for REDUCE/OPT and REDUCE/TRY