Refactor HOST-CONSOLE code, axe `;-- null`, fix input-hook
A goal of having HOST-CONSOLE be written in usermode is that it would
evolve to take advantage of Rebol as a language to express its
state-machine mechanics.  So instead of having the HOST-CONSOLE code
speak directly to the RETURN expectations of the C client, this commit
makes a custom RETURN function that hooks the plain RETURN, and does
various finalization on an instruction that is progressively built
using EMIT commands.

This reduces the amount of code, without changing the workings all
that significantly.  One thing this does do is run the INPUT command
in a sandbox, by offloading it into the INPUT-HOOK console method.
That hook had been broken in a previous commit, but this now makes
the console accrue state line-by-line in an instruction where the
input hook just adds either a STRING! or a BLANK! to the last line
of that accrued multi-line block.

As a pattern, what's happening is that the instructions are made
longer--with fewer dispatches needing to be handled by HOST-CONSOLE
itself.  Debugging becomes more obvious this way, looking at the
instruction stream--which will be useful for console skin developers
to help understand what's going on.

This also gets rid of the `;-- null`, due to the fact that many
functions return null for the express purpose of not giving feedback,
as they do their own printing (HELP, etc.)