0 or 1 length GROUP! evaluator optimization, ANY_INERT()
In SWITCH statements, it was being discussed that literal blocks can
be matched if they are in a GROUP!, e.g.

    data: [print "this just matches"]
    switch data [
        ([print "this just matches"]) [print "this will print"]
    ]

This raised the question of what such evaluations would cost, if on
each switch the group gets evaluated each time.  Really, it need
not cost much of anything...there's no reason to recurse the evaluator
if you know the GROUP! contains *exactly one inert element*.

This does a small reordering of types to make a fast ANY_INERT() test,
and makes 0 element GROUP!s evaluate to null with no recursion, and
inert 1-element GROUP!s evaluate to the item in the group.