Eliminate `<unreachable>` console return status
When the console submitted instructions to C for sandboxing, at times
this instruction would involve a failure/halt/quit that it would be
understood that the next call to HOST-CONSOLE would not (or could not)
recieve a result from that (besides an error).  The hooked RETURN took
the tag `<unreachable>` to indicate this.

However that was a bit confusing, in that the "unreachable" tag result
had to be commented, as to why the next iteration of HOST-CONSOLE
should not expect to get anything back from the instruction.  Since
RETURN was hooked in the first place, this moves the EMIT for the
"unreachable" clause into RETURN itself, which emits it and hence
relieves the callsites of disclaimers of the meaninglessness of what
their argument to RETURN is.