Make KEEP in COLLECT a specialized enclosure of INSERT
With the addition of the /LINE functionality to APPEND+INSERT+CHANGE,
this function did not become added to KEEP.  Rather than add an
extra refinement, this commit makes the intrinsic connection that
KEEP is actually a *specialization* of INSERT for the kept series,
hence inheriting its refinements (e.g. /ONLY, /DUP, /ETC)

However, there's an additional twist, because KEEP is supposed to
return the value passed to it--not the tail of the insertion position
(arguably, the collected-into series is supposed to be a black box
for the COLLECT operation, not exposed during its construction,
which may ultimately present certain optimization opportunities for
the COLLECT abstraction).  Also the series position must be updated
on each insert.  So the specialization is actually of an enclosure
of INSERT which does this work, and all the specialization really
does is remove the series argument.