Make GET return null if variable is not set, /ONLY => /OPT
In R3-Alpha, Rebol2, and Red, getting a variable that is unset is
an error.  Ren-C experimented with the idea that GET would be more
tolerant, and return a blank on failure.

The new model going forward in Ren-C has been to very purposefully use
null results as a kind of "soft" failure.  Most routines reject null
inputs, but if given a blank input will return void.  This makes each
step in an evaluation chain a place that one can elect to put a
TO-VALUE (now, also taking the old name TRY) or not.  Failing to put
in a TRY means one's attention is raised to the failed operation.

So this makes GET return void, while GET/TRY can be used to ask for
blankification.  For ordinary parameters this is the same as TRY GET,
but since blocks cannot hold nulls it's necessary to be a refinement
for block cases:

    >> get [a b] ['asasdfdf 'jklasdf]
    ** error, can't put nulls in block

    >> get/try [a b] ['asasdfdf 'jklasdf]
    [_ _]

The change causes a significant ripple in code dependent on the old
behavior, though little such code should exist in the wild outside
of Ren-C (as R3-Alpha/Rebol2/Red didn't allow GET of unset things)

This also settles on the name /OPT for the refinement to SET that
allows you to pass it NULL.  This solves previous debates and concerns
about what the right name for that refinement is, as /OPT and /TRY
have a coherent meaning.  Little code was affected, as the name
`SET*` was used while anticipating a final decision on the refinement's
name.  It turned out that the missing link was that the name was
different for GET and SET, based on the difference in how the
operations lean.

This required some amount of shakeup in the LOAD code, and the
opportunity was taken to try and make it a bit clearer in the process.