Truncate long outputs, PORT!s
PORT! is a common verbose output result, the dumping of which takes
up too much space on the screen.  R3-Alpha would show nothing, this
changes the default skin to give a small stub to tell you it's a port.

Also includes a truncation so that only the first 20 lines of the first
2048 lines of molded output will be shown.