Add missing blank behavior from last commit, rebUnbox() fix
The previous commit took advantage of a feature in the console where
SKIP of BLANK! would return void.  This dependency was not committed
with the console length-limiting output change, hence the console
was broken for non-void output of less than 20 lines.

Also, rebUnbox() was not properly exported from the API.