Attempt fix for `do <json>` style module loading
The code path for `do <json>` is different from `import <json>`, and
would pass a MODULE! into `load-module`.  This broke several
expectations.  For instance, when LOAD-MODULE gets a FILE! to load
from, it has to load the header...and remember how many lines were
in the header in order to get line numbering correct when loading
the rest of the data.  None of this applies if LOAD-MODULE is
applied to a module which is already loaded.

Exactly why it works this way is nebulous, but clearly LOAD-MODULE is
willing to take a MODULE! as an argument.  While the code is not
mapped out to full understanding, this patches the issue to hopefully
restore at least the prior functionality.

It also removes a couple of CASE/ALL statements, as I consider those
to be further obfuscating.