Force traversal of unfinished material in va_list frames
This commit paves the way for convenience-based features in the
API for automatically releasing values.  The goal is to turn:

    REBVAL *reb_int = rebInteger(c_int);
    REBOOL greater = rebDid(reb_int, "> 304", END);
    rebRelease(reb_int);

Into:

    REBOOL greater = rebDid(rebR(rebInteger(c_int)), "> 304", END);

With the equivalent abbreviation:

    REBOOL greater = rebDid(rebI(c_int), "> 304", END);

The idea is to allow the variadic processing to perform the release
as soon as the handle has been consumed for processing.

However, this requires the variadic feed to be read all the way to
the end...even if it's not going to be run, e.g. in the case of a
failure or throw.  That's a new mechanic which did not exist.  To
get it, places in the source which did not run all the way to the
end of a frame had to be located.

This commit also strengthens the accounting for leaked API handles.
A "leaked API handle" really only gets delegated to the GC, and
is thus not that bad a concern, unless the code running is something
like the console...which is long-running, and in theory may never
exit off the stack, hence accruing handles and keeping a hold on
them indefinitely.  So for the moment, handle leaks are treated as
an error, and caught on the first GC cycle that notices them.

The exception is handles attached to frames which were taken of the
stack due to an error.  These failure frames are considered to be
acceptable places for the system code to allow the GC to pick up
any API handles created during that frame, because cleaning them up
safely in interrupted code would convolute the code too much.

So to get the accounting correct for this commit, rebRescue is also
changed to instantiate a frame, that takes ownership responsibility
for any handles that are created by the C function it is protecting.
It then proxies its result handle back to be owned by whatever frame
called rebRescue (possibly none / top level).

Using this new form of rebRescue() suggested it would be easier to
change the HOST-CONSOLE function to just take one result argument,
that was either NULL, an ERROR!, or the actual result in a 1-element
BLOCK!.  That meant the result could be proxied by rebRescue().