Address three rebRun() issues, mold output in DUMP
Trying several more example cases with rebRun() exposed a couple of
issues.

One is that the bind-while-you-scan implementation would import
variables on demand from lib into user.  But when it did, it would
not transfer the enfix flag (dropped by Move_Value() instructions).
This means if you happened to make the first appearance of an enfix
operator in the user context from a rebRun() call with it as text,
then it would act prefix.

Another was that SET-WORD! and SET-PATH! processing was not preserving
the explicit evaluation flag, so API code which tried to do assignments
in the shallowest level of the variadic would wind up quoting.  This
patches that in a somewhat ad-hoc way, but a more methodical way of
making sure the right flags are inherited in subframes/midframes should
be looked into.

A similar issue was affecting enfix lookback, when the lookback was
into a string portion of a rebRun().

Also, at some point the DUMP command was changed to use PRINT and
did not MOLD GROUP!s being dumped.  This adds the mold, and since
MOLD has a /LIMIT refinement uses that instead of the clip-str tool
that was added there.

evalthing