Eliminate NEW-WORDS extension feature usage in %mod-odbc.c
This makes the ODBC module lean more heavily on libRebol as opposed
to the internal API.  In addition to removing the dependency on the
NEW-WORDS feature of extensions, it also introduces several other
cases of creative usage of the string-based rebRun() interface.

It also removes another deprecated routine, the Get_Typed_Field()
and exports for field selection by name (Get_Field, Sink_Field).
These were not used internally, as it was just a slightly nicer
export of Find_Canon_In_Context.