Eliminate NEW-WORDS extension feature usage in %mod-crypt.c
This commit adds usages of libRebol to get around the need for the use
of the NEW-WORDS feature of extensions.  This means modification to
the Diffie-Hellman private/public key exportation.

While it pushes to use more libRebol, it does not convert the file away
from the internal API entirely.  Some features (such as direct access
of a BINARY!'s data buffer) are not exported in a way that does not
require making a new allocation.  These features are still being
considered and designed.

However, it does go ahead and switch to usage of rebMalloc() when a
data buffer is needed that will ultimately be turned into a BINARY!.