Restore http redirect error functionality
HTTP redirects were never implemented in the HTTP port protocol of
R3-Alpha.  A patch introduced an error message that was sufficient for
clients to implement their own redirects if need be.  This patch had
stopped working due to a recent modification as part of a cleanup.

However, an additional issue is that a code path was becoming exercised
(or an error condition in that code path started triggering) that
brought attention to code from R3-Alpha that would have never worked.
There's a finite list of event types, fixed at compile time.  This is
the list as it was in R3-Alpha, which hasn't really changed:

https://github.com/rebol/rebol/blob/25033f897b2bd466068d7663563cd3ff64740b94/src/boot/sysobj.r#L294

It is not legal to do a `make event! [type: 'xxx ...` unless the xxx
is in that list.  Yet the http protocol would make events with 'error
as the type.  Again, R3-Alpha code:

https://github.com/rebol/rebol/blob/master/src/mezz/prot-http.r#L267

For some reason this began registering as a problem.  This commit adds
`error` to the event type list, which along with fixing the bad `OR`
edit appears to restore the patched redirect behavior.