Manually export host console functions to lib
The whole host startup/console is currently very manually loaded
into its own isolated context by the C startup code.  This way, changes
to functions the console loop depends on (like PRINT or INPUT) that the
user makes will not break the console's functionality.

(It would be better if it used the module system, but R3-Alpha's module
system was largely unfinished, without an obvious story for how it
could be utilized in a C main() function loading code from several
component files.  So when the REPL was changed from being C code to
mostly Rebol, the binding was handled manually as a first step.)

Isolating the module code in its own context meant that functions which
lived with the console code (such as ABOUT or WHY) were no longer
available in the user context.  If it were a module, it would use
the EXPORT feature.  Since it's not, it has to be done manually.

This does that manual exporting for a short list of functions that
live in the console code, as a short-term solution to the problem...
until it is changed to use modules.

Also updates documentation that WHY does not have a ? on the end,
and adds the WHY? version to the r3-legacy list.