Make SWITCH use 0-arity evaluation for match clauses
This commit implements a long-discussed, and fairly significant change,
which makes SWITCH evaluate its match clauses.  As an interim measure,
the version exported to the user context will check to see if you
used a plain WORD! that does not evaluate to a datatype, and warn
that you probably meant to use a LIT-WORD!.  However, in the long term
this warning would be removed.

One major sought-after benefit of this is that types need not be
converted to words or be quoted in order to match:

    switch type of 5 [
        integer! [print "This will work, finally!"]
    ]

However, it carries other benefits, including the ability to mix it
in with ELIDE.

    switch type of x [
        string! [...]
        file! [...]

        elide (
           ...code to run if neither of the above matched...
           ...but continue matching afterward...
           ...other invisibles like DUMP can be used, too...
        )

        block! [...]
        group! [...]
    ]

For reasons of mechanics and clarity, only 0-arity evaluations are
allowed.  On the point of clarity:

    switch 'c [
        third [a b c] [print "Block argument looks like code to run"]
    ]

But on the point of mechanics, it would not be possible to skip over
arbitrary-arity evaluative clauses after a match were found.

A GROUP! may be used to get around this:

    switch 'c [
        (third [a b c]) [print "use a GROUP! for arbitrary code"]
    ]