Change TYPE OF () to be null, not _
In the initial concept of what `type of ()` would give, _ was chosen.
This was a *blank* value, not BLANK! the datatype (you would get BLANK!
the datatype if you did `type of _`.

Consequently it was a bit confusing, leading to the ambiguous statement
"the type is blank".  However, it was chosen this way in order to be
able to use it in switch statements, where literal blanks were legal.

With evaluative switch, null is the better choice for `type of ()`.
Then you can write:

    switch type of :x [
        null [...]
        integer! [...]
        ...
    ]

It still has sort of a weirdness to it--which is unavoidable--to say
the type of null is null, when NULL is not a datatype.  (Remember,
null is defined simply as `null: does []`).  Yet it holds together
better overall.

Accordingly, this makes SWITCH's value to match argument as optional,
and allows the word NULL as one of the exceptions that the interim
safety shim allows (along with words that look up to datatypes)