Add ENTRAP Native
This adds a function which is like TRAP, but puts the result of the
evaluation in a block.  Yet if the result is null, then it returns
null, and if there is a failure raised, it will return that.

Unlike the current implementation of TRAP, if there is a THROW or
RETURN in the body of the code being run, then it will be returned
as a "throw not caught".  This is currently given back as an ERROR!
the same as if a throw hadn't been caught and a failure raised in
the body, though it could be distinguished somehow (e.g. by returning
the not caught throw error in a GROUP!)

As written, it is sufficient to serve as a replacement for the
functionality of %catch-any.r in the tests.