Fix TCC error callback, update primitive bootstrap DUMP
The user natives code was passing a demo value as a test parameter
to the error callback, and asserting the test value was what was
passed to TCC.  But the test was incorrect, leading to an assert.

Also changes the DUMP debug native loaded during boot to do something
more like the higher-level quoting DUMP that loads later.