Remove export of App_Instance from %host-main.c
The Windows build of R3-Alpha used a hidden window for message
processing.  The only use case was asynchronous DNS, which was a
deprecated feature (not being carried forward to IPv6):

https://msdn.microsoft.com/en-us/library/windows/desktop/ms741522(v=vs.85).aspx

One aspect of making this window is that it requires the HINSTANCE of
the application.  That was being passed via a global App_Instance
variable:

    EXTERN_C HINSTANCE App_Instance;  // From Main module.

This complicated linking of libRebol, and since the event strategy is
being rethought this is #ifdef'd out for now.

Long-term, the better way to tunnel such parameters from the host to
extensions would likely be to put a HANDLE! in the environment, and
then those extensions that require the Windows HINSTANCE could
complain if it wasn't there...vs. creating a linker dependency for
all clients.

(Includes fix for stack leak in previous commit related to diagnostics
for user natives.)