Communicate host executable path to console via system/options
While some people believe that argv[0] contains the path to the
running executable for C programs, that is a convention that only
some shells use.  Hence it is up to the host executable to do a
platform-specific call to find out the "true" path:

    https://stackoverflow.com/q/1023306/
    http://stackoverflow.com/a/933996/211160

Over time this had evolved into being done with "host" hook, called by
%host-main.c.  It would then be passed as a parameter to HOST-CONSOLE
on its first run.

This commit changes the process, so that `system/options/boot` is set
through libRebol by the host--or left blank.  The console then
decides if it wants to override that with the contents of argv[0],
which it gets passed as a STRING!.

This is a step toward making the protocol simpler, though hopefully
the code for boot path detection could be made into an extension
which could be reused by other clients.

Also changes the call to HOST-CONSOLE to use ENTRAP instead of a
combination of TRAP and CATCH.