Delete rebPrint/rebBlock/rebFail/rebPanic, add rebJUMPS
A large question about libRebol is "how many API entry points should
there be".  What the goal has been is to *not* try and export C
functions for each operation (rebAppend, rebAppendOnly, rebAppendPart,
rebAppendPartOnly...) but rather to constrain the calls to those that
are needed to express conversions of values from Rebol types to C
types.

The rebFail() and rebPanic() APIs were added as exceptions.  That's
because they were APIs that were used very early on, before string
runs could scan and bind practically.  But also because they did not
return to C, and so having a C entry point with the ATTRIBUTE_NO_RETURN
helped avoid compiler warnings.  Otherwise it would be:

    rebElide("fail", some_failure, END);
    DEAD_END;

However, FAIL and PANIC aren't the only operations that do not return;
THROW and RETURN, etc. don't.  (The API is evolving, and might be able
to actually run RETURN functions that climb out of the C stack, but
this is still an open question.)  Still, any function you know to
invoke a failure could count.

This commit adds a routine tentatively named rebJUMPS, which is a
generic form of rebElide that annotates that the code it is passed
will be jumping.  So it has an ATTRIBUTE_NO_RETURN on it.  This
needed some twisting of the reb-lib.h generation process, to allow
for creating variadic #noreturn inline function wrappers compatible
with the API export process.

In switching the code to use it, this eliminates a lot of redundant
OS_DO_DEVICE error handling code that was added.

It also takes away the rebPrint() and rebBlock() entry points, which
had sketchy semantics--contribution to confusion that, at least for
now, is better handled through the general rebRun()-styled entry
points, rather than introducing more variadic processing.

(Note: The name rebJUMPS is likely to be changed, but just being
tried out for now).