Resolve sporadic GC bug related to device filename results
The current API concept is that all handles start out managed, by the
FRAME! that is in effect when they are created.  If they need to live
longer than that, they must be unmanaged.

The rebRescue() API function was changed so that it would tie any
API handle allocations done in the C function it was wrapping to a
frame that lived as long as the rescue operation.  This allowed clients
of rebRescue to not have to worry about cleaning up handles which
they created in the event of a failure.  However, it also meant that
their handles would expire when the rescue returned.  So the return
result of rebRescue was uniquely "proxied" so that its lifetime would
be tied to the frame that was active when it was called...not the
temporary frame inside the rebRescue.

But rebRescue is now being used by the device layer, so that things at
the "device level" may use API calls and also may fail() with ordinary
Rebol errors.  Yet it might poke some allocated API handles into the
DEVREQ structure which the rebRescue() would not know to proxy, since
it's not part of the return result.

Long term, the general concept is that the information hidden inside
the DEVREQ *would* be the return result, and hence it would be proxied.
However, for now the solution is to unmanage such handles, because
otherwise any evaluator cycle could trigger a recycle and notice the
API handles in the DEVREQ which are tied to an expired frame.