Fix use of freed API handle
Sandboxing code in the console would move an API value into a local
cell, then release the API value.  The intent was to pass the local
value to rebRun().  But it passed the freed API value.

The reason it did this was because it returned a rebBlock(), and
rebBlock could not auto-release.  For the moment, rebBlock() has been
removed...so it's possible to just use rebR() inside a rebRun() that
builds a block via the scanner.