Eliminate OS_ALLOC() in favor of rebAlloc() or malloc()
R3-Alpha had the notion of the host providing a memory allocator and
deallocator.  Code on the host/device side could use this in place of
malloc() and free(), but it was also available for the core to call.
Yet the core had its own specialized allocators and deallocators from
memory pools.

So the concept was that the core would only use the supplied OS_ALLOC()
in cases where it was allocating memory that the device or host
was expected to free.  Similarly, if the host was passing the core
something it should consume, it would pass that memory in as OS_ALLOC()
and the protocol would be that the core would know it could free it.

These memory exchanges were based on the idea that the device layer
did not speak in terms of Rebol types, e.g. REBVAL*.  If it did, then
the way that information could be shared and proxied in an agreed-upon
format would be to use those values.  This is the strategy Ren-C has
been moving toward.

Also in practice, the allocators were really just malloc() and free().
So with their purpose in exchanging freeable memory blocks between
core and host, they had little purpose...except they were wrapped
with macros that made them create the proper type pointers to build
more easily with C++.

In the intervening time, Ren-C got a more useful specialized allocator
in its API, called rebMalloc().  By using a byte-sized series node
underneath the returned memory, it could be reclaimed as a BINARY!
series without another allocation.  That was its initial intended
purpose, but it also had the benefit of being able to be automatically
freed in the event of an error in the frame in which it was allocated.

While currently rebMalloc()'d memory cannot be "unmanaged" except
through conversion to a binary.  Hence when the frame where it is
allocated goes off the stack, that will be considered a leak.  It
thus can't be used as a generic replacement for malloc() (yet), but
most uses of OS_ALLOC() were not persistent...just temporary buffers.
This converts cases which could be converted to a new macro with the
C++ type-safety, rebAlloc() and rebAllocN().  The others it just
makes into malloc() calls for now.

Consequently OS_ALLOC is eliminated, as well as OS_CONFIG, which was
not actually affecting anything.  This further reduces the "hostkit"
hookpoints, with the goal being to get rid of them entirely.