Fix backspace bug in POSIX console, singular array GC
Using backspace in the POSIX console would do a buffer underrun while
searching for the previous UTF8 character.  This wasn't noticed by
memory checking due to the allocator being used for lines, but when
it was changed to use malloc it was caught by address sanitizer.

Also corrects a somewhat esoteric issue which could come up if a GC
occurred on a singular array that itself contained an array with
singular elements.  This could *in theory* overflow the C stack if
a singular array contained a chain of singular arrays longer than
the C stack could be deep.