Add rebHandle() to the libRebol API
The ODBC extension was using rebRun() to assign HANDLE! into objects.
But since there was no rebHandle() function, it was generating the
handle into a temporary DECLARE_LOCAL cell, and then passing that
into rebRun().

However, the local cell was not GC protected.  Hence RECYCLE/TORTURE
would mean the handle's singular array would be freed before it could
get assigned in the Rebol code.

Rather than protect the cell, this continues the goal of making the
ODBC module not use the internal API by adding rebHandle to the
libRebol API.