Distinguish REBARR for UNBOUND/SPECIFIED from EMPTY_ARRAY
To avoid having to check for NULL pointers in the ->binding field, and
go directly to checking node flags, the UNBOUND case does not use
NULL.  It was using the EMPTY_ARRAY, but variadics built from a
MAKE VARIADIC! on a BLOCK! were actually using an ANY-ARRAY! as the
binding, and wanted to distinguish this from the UNBOUND state.
("It is bound, to a content bearing-array, that happens to have been
exhausted and has no more variadic input to offer.")

The simplest resolution to this is just to make the UNBOUND array its
own pointer unique from EMPTY_ARRAY.  This could be made additionally
clear in the future by perhaps putting something in the array, like
an ISSUE! that says it's the unbound/specified signal.