Eliminate rebMold/rebFileToLocal/rebLocalToFile
R3-Alpha had a rather convoluted set of interactions in the OS layer
with functions whose purpose was to transform local filenames to
Rebol filenames, and vice-versa.  Some aspects of this functionality
was exposed with FILE-TO-REBOL and REBOL-TO-FILE.  However, the details
were tied up very much with platform variations in how paths and
slashes were handled--as well as wide characters vs. non-wide.

This was vetted slightly in the Latin1-nowhere change, to the point
where the functions were quarantined a bit into libRebol API variants.
However, they introduced the same problem that adding a C entry point
for every other Rebol function has...needing a wide string variation
vs. not.

In this commit, they are changed to actually use the now-renamed
FILE-TO-LOCAL and LOCAL-TO-FILE API functions.  Within the filesystem
layer on a particular platform, it may wish to implement those APIs
with a C function that is then shared with its export of the
implementation of those two natives...for efficiency.  But for the
moment, this exercises the API.

Also removes rebMold, which was not being used.

Renames the /ONLY refinement on FILE-TO-LOCAL and LOCAL-TO-FILE to
/PASS.