Simplify spelling API, rebString => rebText in API
Early in libRebol, the API "rebSpellingOf" was created.  It could get
the spelling of an ANY-STRING! or ANY-WORD!, and it had variations
that ended in "W" for Wide Char instead of UTF8, and with "Alloc" to
automatically allocate a buffer instead of requiring the caller to
pass one in.

With the evolution of the API, shorter names like rebUnbox, rebElide,
rebRun, rebDid, etc. are being favored for extracting data from
Rebol types.  So for extracting character data, the name "rebSpell"
seems like it might be more succinct.

Also, most clients did not want to get the length returned.  It's also
ambiguous whether the returned length would be in byte counts, length,
or encoding counts (e.g. in UCS2).  This made the API fairly confusing.

With a coming enforcement of no 0 bytes embedded in strings, it would
be purely an optimization to get the length back.  The codepoint count
is available via LENGTH OF in Rebol, and the byte count for UTF8 can
now be acquired via the rebBytes() API...changed to accept strings.

Hence the rebSpellAlloc() interface is changed here to not return
the ambiguous and redundant length.

(An upshot of this is that *in theory* there is no way for a Windows
client to know how many UCS2-encoded WCHARs there are in a string with
the API as written.  In practice, codepoints that don't fit in a wide
character haven't ever been supported...and UCS2 support may not be
a relevant priority going forward.  It could be done in userspace to
make a BINARY! by someone sufficiently motivated, and then extract
those bytes via the API.)