REB_STRING => REB_TEXT, alias TEXT! as STRING! in user context
This makes another long-discussed change, which calls the kind of
ANY-STRING! you make with {braces} or "quotes" known as "TEXT!".

Making a decision is pressed by the need to name functions in the
libRebol API, though also an impending unification of the ANY-WORD!
and ANY-STRING! classes under one umbrella...which is aiming to be
the ANY-STRING! category.

As with the unification of all the ANY-FUNCTION! types under ACTION!,
the effects are not necessarily significant, as the definitions
`string!: text!`, `string?: :text?`, and `to-string: :to-text` cover
most issues of compatibility.

And as with the change from VOID? to NULL?, there are a lot of
comments and secondary namings internal to the codebase that will just
have to get cleaned up over time...though it's not *false* to say
that a TEXT! is a string, since it's in the category.

Rationale for the change is explained here:

https://forum.rebol.info/t/text-vs-string/612