Scan => and =<, kill ?, <- enforce non-NULL, nullable SET
This contains a number of changes provoked by the idea that `SET 'WORD`
should act the same as `WORD:`, while settling on the concept that
`:WORD` acts the same as `GET 'WORD`.

One idea is that the <- operator disallow null.  Its previous main
purpose was precedence manipulation, such that you could write
`return <- if condition [...] else [...]` to stop the default
interpretation as `(return if condition [...]) else [...]`.  But
there's not a particularly compelling reason for it to allow null,
so it can do double duty as a checking construct.

e.g. consider `foo: <- :bar` .  Precedence-wise that's interpreted as
`foo: (:bar)`, which doesn't change anything, since there is no enfix
involved.  But the null check means the person doing the assignment can
be confident that the :bar fetched something.

So this commit makes a greater commitment to the idea that <- (and not
<| as in F# or Elm, and not $ as in Haskell) is the means of making
an implicit GROUP!.  With that commitment, -> is taken away from
lambda, which is shifted to the heavier-looking =>.  This required
a change to the scanner, as => and =< were not allowed previously
(despite being added to root words in lib long ago, the scanner still
wouldn't allow them).

(Hence the mezzanine must use the non-enfix form LAMBDA, as => cannot
be scanned in bootstrap.)

Given the decision that <- will be the means for null checking, SET
and SET-WORD! are reconciled with the elimination of /OPT refinement,
and hence SET* is not needed to unset a variable.  `set 'x ()` works,
and those concerned are to use `set 'x <- ...`

An unused definition of ? is killed as well, conversation here:

https://forum.rebol.info/t/the-question-of-plain-old-and-plain-old/624

Since =< is now legal, it is halfheartedly promoted as the "main"
form of comparison by changing the native LESSER-OR-EQUAL? to
EQUAL-OR-LESSER?, while preserving <= and LESSER-OR-EQUAL? as
mezzanine synonyms.

An oversight in the console in delivering internal errors is corrected,
as a print command was issued in generated code after a fail was
issued, which meant it would never be seen.